feat: differentiation
