+++ src/contentscript/contentscript.scss
@@ -1,22 +1,42 @@
-.static-imgage {
-  display: none;
-}
-
-img.boxart-image  {
-  display: none;
-}
+/* thumbanil */
 .boxart-container {
   border: 1px solid white;
   display: flex;
   flex-direction: column;
 }
+
+.fallback-text {
+  display: none;
+}
+
+.video-title {
+  -webkit-box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  box-sizing: border-box;
+  position: absolute;
+  bottom: 0;
+  left: 8%;
+  right: 8%;
+  margin: 0;
+  padding: 0 0 10%;
+  text-align: center;
+  font-size: 1.5em;
+  font-weight: bold;
+  white-space: nowrap;
+  overflow: hidden;
+  -o-text-overflow: ellipsis;
+  text-overflow: ellipsis;
+}
+
+/* background image */
+.static-imgage {
+  display: none;
+}
+
 .spotlight-background-fill {
   display: none;
 }
 .title .logo {
-  display: none;
-}
-.VideoContainer video {
   display: none;
 }
 .video-component-container {
@@ -29,7 +49,7 @@ img.boxart-image  {
   }
 }
 .billboard-title:before {
-  content: "Netflix";
+  content: "Unknown";
   display: block;
   font-size: 1.7vw;
 }
@@ -38,10 +58,4 @@ img.boxart-image  {
   margin: 0 auto;
   font-size: 1.0vw;
   text-align: center; 
-}
-.fallback-text-container {
-  background-color: white;
-}
-.static-image {
-  display: none;
 }
