+++ src/contentscript/interface.ts
@@ -1,4 +1,5 @@
 export interface configTargets {
   actors: string;
   titles: string;
+  headings: string;
 }
